requirements.txt
click>=8.0.0
rich>=13.0.0
requests>=2.28.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
pydantic>=2.0.0
github-copilot-sdk>=0.1.0
